feat: add route and model for person
Using express.Router a new route (/persons) was created. It is good
practice to have separate entities live inside separate files.

A new schema using mongoose.Schema was created. The reference for the
schema was taken from the idea in the Readme file.

Added endpoint for get request, which fetches data from the person db.